Fixed Bug that an error os thrown when the app request permissions because the Android Activity is not build yet. (#22)
--> Refactored that the navigation is now an extra page

--> MyApp is now a Future Builder that first checks the Permissions and then fully Builds the App
