+++ lib/main.dart
@@ -1,19 +1,16 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:jog_dog/pages/page_history.dart';
-import 'package:jog_dog/pages/page_home.dart';
-import 'package:jog_dog/pages/page_settings.dart';
 import 'package:jog_dog/theme/theme.dart';
 import 'package:jog_dog/utilities/debug_logger.dart';
 import 'package:jog_dog/utilities/session_manager.dart';
 import 'package:permission_handler/permission_handler.dart';
 
+import 'pages/page_navigation.dart';
+
 var logger;
 
 void main() {
   runApp(const MyApp());
-  requestPermissions();
   SessionManager().loadSessionsFromJson();
   // If in Debug Mode this Code will be executed
   // Else this code will be removed automatically
@@ -26,64 +23,31 @@ void main() {
   }
 }
 
-class MyApp extends StatefulWidget {
+class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
   @override
-  State<MyApp> createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  late int currentPageIndex = 0;
-
-  @override
   Widget build(BuildContext context) {
-    SystemChrome.setPreferredOrientations([
-      DeviceOrientation.portraitUp,
-      DeviceOrientation.portraitDown,
-    ]);
-    return MaterialApp(
-      theme: AppTheme.lightTheme,
-      darkTheme: AppTheme.darkTheme,
-      themeMode: ThemeMode.system,
-      home: Scaffold(
-        bottomNavigationBar: NavigationBar(
-          selectedIndex: currentPageIndex,
-          destinations: const [
-            NavigationDestination(
-                icon: Icon(Icons.house_outlined), label: "Home"),
-            NavigationDestination(icon: Icon(Icons.history), label: "History"),
-            NavigationDestination(icon: Icon(Icons.settings), label: 'Settings')
-          ],
-          onDestinationSelected: (int index) {
-            setState(() {
-              currentPageIndex = index;
-            });
-          },
-        ),
-        body: <Widget>[
-          const Home(),
-          const History(),
-          const Settings()
-        ][currentPageIndex],
-      ),
+    return FutureBuilder(
+      future: requestPermissions(),
+      builder: (BuildContext context, AsyncSnapshot<void> snapshot) {
+        return MaterialApp(
+          theme: AppTheme.lightTheme,
+          darkTheme: AppTheme.darkTheme,
+          themeMode: ThemeMode.system,
+          home: const NavigationPage(),
+        );
+      },
     );
   }
 }
 
-Future<bool> requestPermissions() async {
-  bool reqSuc = false;
-
-  if (await Permission.location.serviceStatus.isEnabled) {
-    if (kDebugMode) {
-      logger.i("Permission: location already granted");
-    }
-    return true;
-  } else if (await Permission.location.serviceStatus.isDisabled) {
+Future<void> requestPermissions() async {
+  if (await Permission.location.serviceStatus.isDisabled) {
     if (kDebugMode) {
       logger.e("No Location Service Found!");
     }
-    return false;
+    return;
   }
 
   // All needed Permissions should be stored in this map and will be
@@ -91,7 +55,6 @@ Future<bool> requestPermissions() async {
   Map<Permission, PermissionStatus> allPermissionsStatus = await [
     Permission.location,
     Permission.storage,
-    //Permission.activityRecognition,
   ].request();
 
   allPermissionsStatus.forEach(
@@ -100,20 +63,17 @@ Future<bool> requestPermissions() async {
         if (kDebugMode) {
           logger.i("Permission: $key already granted");
         }
-        reqSuc = true;
       } else if (currentStatus.isDenied) {
         //currentStatus = await key.request();
         if (kDebugMode) {
           logger.i("Permission: $key was not acceppted!");
         }
-        reqSuc = true;
       } else if (currentStatus.isPermanentlyDenied) {
         if (kDebugMode) {
           logger.i("Permission: $key is permanently denied");
         }
         openAppSettings();
         if (currentStatus.isGranted) {
-          reqSuc = true;
         } else {
           if (kDebugMode) {
             logger.i("Permission: $key still denied");
@@ -122,6 +82,4 @@ Future<bool> requestPermissions() async {
       }
     },
   );
-
-  return reqSuc;
 }
