+++ lib/pages/page_navigation.dart
@@ -1,0 +1,46 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
+import 'page_history.dart';
+import 'page_home.dart';
+import 'page_settings.dart';
+
+class NavigationPage extends StatefulWidget {
+  const NavigationPage({super.key});
+
+  @override
+  State<NavigationPage> createState() => _NavigationPage();
+}
+
+class _NavigationPage extends State<NavigationPage> {
+  late int currentPageIndex = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    SystemChrome.setPreferredOrientations([
+      DeviceOrientation.portraitUp,
+      DeviceOrientation.portraitDown,
+    ]);
+    return Scaffold(
+      bottomNavigationBar: NavigationBar(
+        selectedIndex: currentPageIndex,
+        destinations: const [
+          NavigationDestination(
+              icon: Icon(Icons.house_outlined), label: "Home"),
+          NavigationDestination(icon: Icon(Icons.history), label: "History"),
+          NavigationDestination(icon: Icon(Icons.settings), label: 'Settings')
+        ],
+        onDestinationSelected: (int index) {
+          setState(() {
+            currentPageIndex = index;
+          });
+        },
+      ),
+      body: <Widget>[
+        const Home(),
+        const History(),
+        const Settings()
+      ][currentPageIndex],
+    );
+  }
+}
